fix: resolve CI failures — clippy await_holding_lock and readability ratchet
- os_apps.rs: restructure test to fetch Turso data before acquiring the
  RwLock write guard, avoiding holding it across an await point (fixes
  clippy::await_holding_lock with -D warnings)
- actor.rs: extract #[cfg(test)] mod tests to actor_test.rs, reducing
  production file from 1013 → 742 lines (keeps PROD_FILES_GT1000=0 and
  PROD_MAX_FILE_LINES within the 987 baseline)
