+++ crates/temper-server/src/entity_actor/actor_test.rs
@@ -1,0 +1,271 @@
+use super::*;
+use std::time::Duration;
+use temper_jit::table::TransitionTable;
+use temper_runtime::ActorSystem;
+
+const ORDER_IOA: &str = include_str!("../../../../test-fixtures/specs/order.ioa.toml");
+
+fn order_table() -> Arc<RwLock<TransitionTable>> {
+    Arc::new(RwLock::new(TransitionTable::from_ioa_source(ORDER_IOA)))
+}
+
+// =============================================
+// DST-FIRST: Test the actor through the runtime
+// =============================================
+
+#[tokio::test]
+async fn dst_entity_starts_in_initial_state() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-1", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-1");
+
+    let response: EntityResponse = actor_ref
+        .ask(EntityMsg::GetState, Duration::from_secs(1))
+        .await
+        .unwrap();
+
+    assert_eq!(response.state.status, "Draft");
+    assert_eq!(response.state.entity_id, "order-1");
+    assert_eq!(response.state.item_count, 0);
+    assert!(response.state.events.is_empty());
+}
+
+#[tokio::test]
+async fn dst_add_item_then_submit() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-2", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-2");
+
+    // Add an item (Draft -> Draft, item_count 0 -> 1)
+    let r: EntityResponse = actor_ref
+        .ask(
+            EntityMsg::Action {
+                name: "AddItem".into(),
+                params: serde_json::json!({"ProductId": "prod-1"}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+    assert!(r.success);
+    assert_eq!(r.state.status, "Draft");
+    assert_eq!(r.state.item_count, 1);
+
+    // Submit (Draft -> Submitted)
+    let r: EntityResponse = actor_ref
+        .ask(
+            EntityMsg::Action {
+                name: "SubmitOrder".into(),
+                params: serde_json::json!({"ShippingAddressId": "addr-1"}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+    assert!(r.success, "submit should succeed, got: {:?}", r.error);
+    assert_eq!(r.state.status, "Submitted");
+    assert_eq!(r.state.events.len(), 2); // AddItem + SubmitOrder
+}
+
+#[tokio::test]
+async fn dst_cannot_submit_without_items() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-3", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-3");
+
+    // Try to submit with 0 items -- should fail
+    let r: EntityResponse = actor_ref
+        .ask(
+            EntityMsg::Action {
+                name: "SubmitOrder".into(),
+                params: serde_json::json!({}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+    assert!(!r.success);
+    assert_eq!(r.state.status, "Draft"); // Still in Draft
+}
+
+#[tokio::test]
+async fn dst_full_order_lifecycle() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-4", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-4");
+
+    // Draft -> AddItem -> SubmitOrder -> ConfirmOrder -> ProcessOrder -> ShipOrder -> DeliverOrder
+    let actions = [
+        ("AddItem", serde_json::json!({})),
+        ("SubmitOrder", serde_json::json!({})),
+        ("ConfirmOrder", serde_json::json!({})),
+        ("ProcessOrder", serde_json::json!({})),
+        ("ShipOrder", serde_json::json!({})),
+        ("DeliverOrder", serde_json::json!({})),
+    ];
+
+    let expected_states = [
+        "Draft",      // after AddItem
+        "Submitted",  // after SubmitOrder
+        "Confirmed",  // after ConfirmOrder
+        "Processing", // after ProcessOrder
+        "Shipped",    // after ShipOrder
+        "Delivered",  // after DeliverOrder
+    ];
+
+    for (i, (action, params)) in actions.into_iter().enumerate() {
+        let r: EntityResponse = actor_ref
+            .ask(
+                EntityMsg::Action {
+                    name: action.into(),
+                    params,
+                    cross_entity_booleans: std::collections::BTreeMap::new(),
+                },
+                Duration::from_secs(1),
+            )
+            .await
+            .unwrap();
+        assert!(r.success, "step {i} ({action}) failed: {:?}", r.error);
+        assert_eq!(
+            r.state.status, expected_states[i],
+            "step {i} ({action}) wrong state"
+        );
+    }
+
+    // Verify full event log
+    let r: EntityResponse = actor_ref
+        .ask(EntityMsg::GetState, Duration::from_secs(1))
+        .await
+        .unwrap();
+    assert_eq!(r.state.events.len(), 6);
+    assert_eq!(r.state.status, "Delivered");
+}
+
+#[tokio::test]
+async fn dst_cancel_from_draft() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-5", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-5");
+
+    let r: EntityResponse = actor_ref
+        .ask(
+            EntityMsg::Action {
+                name: "CancelOrder".into(),
+                params: serde_json::json!({"Reason": "changed mind"}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+    assert!(r.success);
+    assert_eq!(r.state.status, "Cancelled");
+}
+
+#[tokio::test]
+async fn dst_cannot_cancel_shipped_order() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+    let actor = EntityActor::new("Order", "order-6", table, serde_json::json!({}));
+    let actor_ref = system.spawn(actor, "order-6");
+
+    // Drive to Shipped
+    for action in &[
+        "AddItem",
+        "SubmitOrder",
+        "ConfirmOrder",
+        "ProcessOrder",
+        "ShipOrder",
+    ] {
+        let _: EntityResponse = actor_ref
+            .ask(
+                EntityMsg::Action {
+                    name: action.to_string(),
+                    params: serde_json::json!({}),
+                    cross_entity_booleans: std::collections::BTreeMap::new(),
+                },
+                Duration::from_secs(1),
+            )
+            .await
+            .unwrap();
+    }
+
+    // Try to cancel -- should fail
+    let r: EntityResponse = actor_ref
+        .ask(
+            EntityMsg::Action {
+                name: "CancelOrder".into(),
+                params: serde_json::json!({}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+    assert!(!r.success);
+    assert_eq!(r.state.status, "Shipped"); // Still Shipped
+    assert!(r.error.unwrap().contains("not valid"));
+}
+
+#[tokio::test]
+async fn dst_multiple_actors_independent() {
+    let system = ActorSystem::new("dst");
+    let table = order_table();
+
+    let a1 = system.spawn(
+        EntityActor::new("Order", "order-A", table.clone(), serde_json::json!({})),
+        "order-A",
+    );
+    let a2 = system.spawn(
+        EntityActor::new("Order", "order-B", table.clone(), serde_json::json!({})),
+        "order-B",
+    );
+
+    // Cancel order A
+    let _: EntityResponse = a1
+        .ask(
+            EntityMsg::Action {
+                name: "CancelOrder".into(),
+                params: serde_json::json!({}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+
+    // Add item to order B
+    let _: EntityResponse = a2
+        .ask(
+            EntityMsg::Action {
+                name: "AddItem".into(),
+                params: serde_json::json!({}),
+                cross_entity_booleans: std::collections::BTreeMap::new(),
+            },
+            Duration::from_secs(1),
+        )
+        .await
+        .unwrap();
+
+    // Verify independence
+    let r1: EntityResponse = a1
+        .ask(EntityMsg::GetState, Duration::from_secs(1))
+        .await
+        .unwrap();
+    let r2: EntityResponse = a2
+        .ask(EntityMsg::GetState, Duration::from_secs(1))
+        .await
+        .unwrap();
+
+    assert_eq!(r1.state.status, "Cancelled");
+    assert_eq!(r2.state.status, "Draft");
+    assert_eq!(r2.state.item_count, 1);
+}
